requirements/base.txt
Django==3.2
python-decouple==3.6
psycopg2-binary==2.9.1
django-ninja==0.18.0
djangorestframework-jwt==1.11.0
djangorestframework==3.12.4
djangorestframework-stubs==1.4.0
